change function name from `to_piet` to `draw`
I'm kind of conflicted about this change.
On the one hand, `to_piet` is, I think,
a weak description of what this function does.
On the other hand changing a function name
because I don't like it is dubious reasoning.

I think `to` implies some kind of conversion,
and, from what I can tell, `piet` is mostly
incedental in this function.

`pub` was also added to this function to
expose it to other app developers and
library writers to create custom widgets
based on `ImageData`.